+++ librosa/decompose.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-"""Spectrogram decomposition
+"""
+Spectrogram decomposition
 =========================
 .. autosummary::
     :toctree: generated/